Add verification for type V2
