+++ auto-agents-framework/dsn-kol-schemas.json
@@ -1,0 +1,204 @@
+{
+  "response": {
+    "type": "object",
+    "properties": {
+      "type": {
+        "type": "string",
+        "const": "response"
+      },
+      "tweet": {
+        "type": "object",
+        "properties": {
+          "id": {
+            "type": "string"
+          },
+          "text": {
+            "type": "string"
+          },
+          "username": {
+            "type": "string"
+          },
+          "timeParsed": {
+            "type": "string"
+          },
+          "thread": {
+            "type": "array",
+            "items": {
+              "$ref": "#/properties/tweet"
+            }
+          }
+        },
+        "required": [
+          "id",
+          "text",
+          "username",
+          "timeParsed"
+        ],
+        "additionalProperties": false
+      },
+      "decision": {
+        "type": "object",
+        "properties": {
+          "shouldEngage": {
+            "type": "boolean"
+          },
+          "reason": {
+            "type": "string"
+          }
+        },
+        "required": [
+          "shouldEngage"
+        ],
+        "additionalProperties": false
+      },
+      "content": {
+        "type": "string",
+        "description": "The response to the tweet"
+      },
+      "strategy": {
+        "type": "string",
+        "description": "The strategy used to generate the response"
+      },
+      "previousCid": {
+        "type": "string"
+      },
+      "signature": {
+        "type": "string"
+      },
+      "timestamp": {
+        "type": "string"
+      },
+      "agentVersion": {
+        "type": "string"
+      }
+    },
+    "required": [
+      "type",
+      "tweet",
+      "decision",
+      "content",
+      "strategy",
+      "signature",
+      "timestamp",
+      "agentVersion"
+    ],
+    "additionalProperties": false,
+    "$schema": "http://json-schema.org/draft-07/schema#"
+  },
+  "skipped_engagement": {
+    "type": "object",
+    "properties": {
+      "type": {
+        "type": "string",
+        "const": "skipped"
+      },
+      "tweet": {
+        "type": "object",
+        "properties": {
+          "id": {
+            "type": "string"
+          },
+          "text": {
+            "type": "string"
+          },
+          "username": {
+            "type": "string"
+          },
+          "timeParsed": {
+            "type": "string"
+          },
+          "thread": {
+            "type": "array",
+            "items": {
+              "$ref": "#/properties/tweet"
+            }
+          }
+        },
+        "required": [
+          "id",
+          "text",
+          "username",
+          "timeParsed"
+        ],
+        "additionalProperties": false
+      },
+      "decision": {
+        "type": "object",
+        "properties": {
+          "shouldEngage": {
+            "type": "boolean"
+          },
+          "reason": {
+            "type": "string"
+          }
+        },
+        "required": [
+          "shouldEngage"
+        ],
+        "additionalProperties": false
+      },
+      "previousCid": {
+        "type": "string"
+      },
+      "signature": {
+        "type": "string"
+      },
+      "timestamp": {
+        "type": "string"
+      },
+      "agentVersion": {
+        "type": "string"
+      }
+    },
+    "required": [
+      "type",
+      "tweet",
+      "decision",
+      "signature",
+      "timestamp",
+      "agentVersion"
+    ],
+    "additionalProperties": false,
+    "$schema": "http://json-schema.org/draft-07/schema#"
+  },
+  "generated_tweet": {
+    "type": "object",
+    "properties": {
+      "type": {
+        "type": "string",
+        "const": "posted"
+      },
+      "content": {
+        "type": "string"
+      },
+      "tweetId": {
+        "type": [
+          "string",
+          "null"
+        ]
+      },
+      "previousCid": {
+        "type": "string"
+      },
+      "signature": {
+        "type": "string"
+      },
+      "timestamp": {
+        "type": "string"
+      },
+      "agentVersion": {
+        "type": "string"
+      }
+    },
+    "required": [
+      "type",
+      "content",
+      "tweetId",
+      "signature",
+      "timestamp",
+      "agentVersion"
+    ],
+    "additionalProperties": false,
+    "$schema": "http://json-schema.org/draft-07/schema#"
+  }
+}
